alphaevolve_core/src/evaluation/sandbox.py
import docker
import os
import tempfile
from typing import Dict, Any, Tuple

class DockerSandbox:
    def __init__(self, image_name="python:3.9-slim"): # Using a slim Python image
        self.client = docker.from_env()
        self.image_name = image_name
        # TODO: Build the Docker image if it doesn't exist

    def run_code(self, code: str, function_name: str, test_input: Any, timeout: int = 10) -> Tuple[Any, str, str]:
        """
        Runs the provided Python code in an isolated Docker container.

        Args:
            code: The Python code to run.
            function_name: The name of the function to call in the code.
            test_input: The input to pass to the function.
            timeout: Maximum execution time in seconds.

        Returns:
            A tuple containing:
                - result: The output of the function execution.
                - stdout: The standard output from the container.
                - stderr: The standard error from the container.
        """
        # This is a placeholder implementation.
        # The actual implementation will involve writing code to a temp file,
        # creating a runner script, mounting them into a container, and running it.
        print("Running code in Docker sandbox (placeholder)...")
        print(f"Code:\n{code}")
        print(f"Function: {function_name}")
        print(f"Input: {test_input}")

        # Dummy results
        result = None
        stdout = "Simulated stdout\n"
        stderr = ""

        try:
            # Simulate execution
            import time
            time.sleep(1) # Simulate some work
            result = f"Simulated result for input {test_input}"
        except Exception as e:
            stderr = f"Simulated error: {e}"

        return result, stdout, stderr

    def __del__(self):
        # Clean up resources if necessary
        pass
