alphaevolve_core/requirements.txt
google-generativeai
python-dotenv
docker-py
